Backend/requirements.txt
fastapi
uvicorn
web3
python-dotenv
cryptography
sqlalchemy
psycopg2
redis
bcrypt
PyJWT
yagmail
pyyaml
ua-parser
user-agents
passlib
qrcode[pil]
reportlab
py-solc-x




